Version 0.3.2 - Signals and Trading
Create: Functions that give indicator signals were written. A test algorithm that buys and sells according to the RSI Indicator was written.
Update: Sales transactions have been added to the DCA Algorithm. The test algorithms were updated so that transactions are made from the current candle - 1, not from the current candle. All pages are linked to each other in the import section of the pages. Some imports have been renamed.
Bug Fix: A solution has been applied to the error in the EMA indicator function.