Version 0.3.5
Create: The file of the names of the libraries was created.
Update: Library links have been moved to library.py. The coinlist.csv file has been replaced with coinlist.txt. The business.py file has been moved to the dataops.py file in the engine. The backtest.py file has been moved to the algotest.py file in the engine.
Delete: Some indicators and default settings were moved to the comment line for a short time. (It will be deleted in the next version and will not be needed in version 1.0.0.)