Version 0.5.0
Import: TeleBot library and RE library added.

Create: List.WriteCoin function has been written and you can now get a Coin List via Telegram. List.WriteFullCoin and List.WriteFullCoin functions have been written so that when the user enters a coin, it will be possible to query whether this coin is listed on Binance or not. The Message.Start function has been created so that the Telegram bot can be activated. The Message.Get function was created so that messages from Telegram can be read and processed. Thanks to the Algorithm.FullTest function, backtests of the coin purchased from Telegram will be made. Thanks to the Calculate.FindCoin function, if the coin is not found an error message will be returned, if it is found it will return the value True. Calculate.GetChangeWallet function was written, this function is used to calculate the daily changes of the wallet. If an analysis is requested for the coinlist.txt file from Telegram with the Analysis.CoinListInfo function, the minimum and maximum rising coins can be listed. Thanks to the Analysis.WalletInfo function, the total value, spot value and earn value of the wallet can be listed. Thanks to Analysis.CoinInfo, if the user wants a coin to be analyzed via Telegram, it will work and give the buy and sell signals of some indicators to the user.

Update: Files have been renamed. The DATETIME query has been removed from the Candle.Write function. List.ReadCoin function has been updated, if the coinlist.txt file is not found an error message will be returned. List.WriteChange function has been updated. Now, if the entered coins do not actually exist, the error message will return and ignore that coin. If the List.ReadChange function cannot find the csv file, it will run the List.WriteChange function. Thanks to the update in the Wallet.Write function, only coins worth more than 1 dollar will now be written.

Delete: Algorithm.Algorithms that backtested with some indicators such as DCA, EMA, RSI, StochRSI were deleted. It will not be used in this version.

Bug Fix: Although the balance[i] value was not used in the Calculate.TotalWallet function, it was not deleted, which gave an error. Replaced by USDTBalance[i].

Optimization: Indicator.SMA, EMA, RSI, StochRSI functions have been reduced to a single line.